requirement.txt
openai
streamlit
langchain-openai
python-dotenv # load all env varaible
pandas
beautifulsoup4 # for web scraping
termcolor # for coloring the texts

# PDF query
PyPDF2
# FAISS (Facebook AI Similarity Search) helps in fast nearest neighbor search in high-dimensional spaces.
faiss-cpu
# tiktoken helps in creating token
tiktoken
huggingface_hub
langchain-huggingface

#Embeddings
langchain_chroma
sentence_transformers


